# Use map in stringer