* master
    ```
    goos: linux
    goarch: amd64
    pkg: github.com/at-wat/ebml-go
    BenchmarkMarshal-4     	10000000	      5248 ns/op	    1728 B/op	      51 allocs/op
    BenchmarkUnmarshal-4   	10000000	      4261 ns/op	     360 B/op	      43 allocs/op
    ```

* this
    ```
    goos: linux
    goarch: amd64
    pkg: github.com/at-wat/ebml-go
    BenchmarkMarshal-4     	10000000	      5340 ns/op	    1728 B/op	      51 allocs/op
    BenchmarkUnmarshal-4   	10000000	      4311 ns/op	     360 B/op	      43 allocs/op
    ```